Add comment about commented out code in fixed code
The code in question was to determine if the files are identical. It
won't compile on modern systems and I'm much too lazy to fix it. But
rather than just comment it out I think it's good to show that it was
part of the code originally and did work.

Other things I changed to get it to compile I didn't comment but one can
just look at the diff from the two compilable entries and the original
file for that.